r/common: add PF Reloader script
pf_reloader is an rc script that reloads the pf ruleset once the DNS
subsystem has started. If pf.conf has a DNS name in it, and the DNS service
is not enabled, **pf will fail to load, and will proceed to block all
traffic.** Thus, by reloading PF once DNS is up, the ruleset will be
loaded correctly.